pyproject.toml
[tool.poetry]
name = "unet"
version = "0.1.0"
description = "UNet implementation in TensorFlow"
license = "Apache-2.0"

authors = ["Victor I. Afolabi <javafolabi@gmail.com>"]
maintainers = [
  "Victor I. Afolabi <javafolabi@gmail.com>",
]

readme = "README.md"
repository = "https://github.com/victor-iyi/unet"
keywords = ["tensorflow", "machine learning", "deep learning"]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache 2.0",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
  { include = "unet" },
]

[tool.poetry.dependencies]
python = "^3.9"

# TensorFlow datasets.
tensorflow-datasets = "*"
# TensorFlow for MacOS
tensorflow-macos = "*"
# TensoorFlow Metal acceleration
tensorflow-metal = "*"
# Data visulaization.
matplotlib = "*"
# Model summary plot [optional]
pydot = "*"

[tool.poetry.dev-dependencies]
# Automatically format Python code to PEP 8 style guide.
autopep8 = "*"
# A coverage plugin to provide sensible default settings.
covdefaults = "^2.2.0"
# The modular source checker.
flake8 = "*"
# Optional static typing for Python.
mypy = "^0.982"
# REPL editor environment for data science.
# juypterlab = "^3.4.8"
# Managing and maintaining multi-language pre-commit hooks.
pre-commit = "^2.20.0"
# Python style guide checker.
pydocstyle = "^6.1.1"
# Simple powerful testing with Python.
pytest = "^7.1.3"
# Python plugin for measuring coverage.
pytest-cov = "^4.0.0"
# Generic viirtualenv management and test cli tool.
tox = "^3.26.0"


[tool.poetry.scripts]
unet = "main:main"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/victor-iyi/unet/issues"

# Poetry build system. See: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "unet"
check_untyped_defs = true
ignore_missing_imports = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.coverage.run]
plugin = "covdefaults"
